add foundation for mysql "emergency password reset"
Adding a feature which asks the user if they want to perform an emergency password reset on root if they guess the password incorrectly 3 times.

The method used to perform the reset will kill the mysql process, edit /etc/mysql/my.cnf and attempt to reset the password with "skip-grant-tables" enabled.